pyproject.toml
[project]
name = "src"
version = "0.0.1"
description = "SfM with multi-view 3D priors"
authors = [
  {name = "Daniel Korth", email = "daniel.korth@tum.de"}
]
dependencies = [
  # Core ML/CV
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "lightning>=2.0.0",
  "torchmetrics>=0.11.4",

  # Hydra ecosystem
  "hydra-core==1.3.2",
  "hydra-colorlog==1.2.0",
  "hydra-optuna-sweeper==1.2.0",
  "omegaconf",

  # Utilities
  "rootutils",
  "rich>=13.0.0",
  "wandb",
  "tqdm",
  "numpy",
  "Pillow",  # PIL
  "pandas",  # for data analysis

  # DUSt3R dependencies
  "roma",
  "einops",
  "trimesh",
  "huggingface-hub[torch]>=0.22",

  # Visualization
  "plotly",
  "nbformat",
  "torchtyping",
  "matplotlib",
  "tensorboard",
  "pyglet<2",
  "rerun-sdk",  # for visualization
  "open3d",

  # OpenCV
  "opencv-python",

  # Scientific computing
  "scipy",

  # COLMAP
  "pycolmap",

  # Type checking
  "jaxtyping",  # for type hints with torch tensors
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pre-commit>=3.0.0",
  "black",
  "isort",
  "lightning-utilities",  # for rank_zero_only decorator
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
